requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1.2
django-enumfields2==3.0.2
django-environ==0.11.2
django-filter==24.3
django-recurrence==1.11.1
django-rest-auth==0.9.5
djangorestframework==3.15.2
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
uritemplate==4.1.1
